requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
asyncio==3.4.3
attrs==23.2.0
bidict==0.23.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
discord==2.3.2
discord.py==2.3.2
disnake==2.9.1
distlib==0.3.8
filelock==3.14.0
frozenlist==1.4.1
h11==0.14.0
idna==3.7
multidict==6.0.5
packaging==24.0
platformdirs==4.2.1
pycparser==2.22
pykuma==1.0.4
PyNaCl==1.5.0
python-engineio==4.9.0
python-socketio==5.11.2
requests==2.31.0
simple-websocket==1.0.0
uptime_kuma_api==1.2.1
urllib3==2.2.1
virtualenv==20.26.1
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4
